MLTB: Async Telegram bot for monitoring devices in wifi (linksys)
This bot is designed to monitor devices in the wifi network for linksys
rotuters. A vulnerability in linksys is used to obtain information about
devices in the network, so there's no need to even login to the router,
it's enough just to be in the same network as the router.

Supported commands:
Help:
d, devices - Get current devices list
r, register - Register device changes notification
u, unregister - Unregister device changes notification
h, help - Command list

The list of commands is easy (supposedly) to expand if necessary. To do this,
simply expand TelegramCommands.list with a new command.